DA preparation: surface observation operator modified
FossilOrigin-Name: c00bf6f3f7cdf12046aa702f804a701aae10f0b649ad5ba319cee7a843a820cd

+++ MSM-Tactical/sys/ush/raddprtbbase.sh
@@ -12,7 +12,6 @@ QADJ=${QADJ:-no} #super saturation and dry adjustment
 BP=${BP} #with boundary perturbation
 MSMDIR=/home/nakashita/Development/grmsm/MSM-Tactical
 SRCDIR=${MSMDIR}/usr/post
-head=${HEAD:-bv$TETYPE}
 DATADIR=${RUNDIR:-/zdata/grmsm/work/dpac/rsm27/$SDATE}
 BASE0=${BASEDIR0:-/zdata/grmsm/work/gfsp2rsm27_nomad}
 BASE1=${BASEDIR1:-/zdata/grmsm/work/gfsp2rsm27_rda}
@@ -41,6 +40,15 @@ if [ ! -d $BASEDIR ];then
   fi
 fi
 echo $BASEDIR
+#restart check
+MEM=$MEMBER
+PMEM=`printf '%0.3d' $MEM` #prtb member
+OUTDIR=$BASEDIR/${PMEM}
+if [ -d $OUTDIR ]; then
+  echo 'base perturbation already done.'
+  exit
+fi
+
 rm -rf tmp
 mkdir -p tmp
 cd tmp
